requirements.txt
# Core dependencies for trademark extraction system
fastapi
uvicorn[standard]
python-multipart

# LangChain and AI dependencies
langchain-groq
langchain-core

# PDF processing
PyMuPDF

# Additional utilities
python-dotenv
requests

# Optional: For enhanced functionality
pillow
numpy

# Trademark comparison dependencies
jellyfish
fuzzywuzzy
python-Levenshtein

# Streamlit UI dependencies
streamlit
plotly
pandas

# Development dependencies (optional)
pytest
pytest-asyncio 